Add files via upload
Hi guys, I produced a map of sightings by state. More red color indicates a higher number of sightings. There is still much to add to this map but feel this is a good start. Please review the README.md file before running this code. To get this map to work you only need to worry about files contained in this directory (State_Sightings_Map/).